backend/requirements.txt
fastapi
uvicorn
python-multipart
python-pptx
requests
python-dotenv
pytest
pdf2image
